接口2.5 2.6

+++ src/main/java/com/parkinglot_backend/entity/Voucher.java
@@ -23,7 +23,7 @@ public class Voucher implements Serializable {
 
     private String rules;
 
-    private Integer payvalue;
+    private Double payvalue;
 
     private Integer actualvalue;
